MinMax values for exposure,..
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@9355 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/focusd/optec.cpp
@@ -52,8 +52,8 @@ class Optec:public Focusd
 		virtual int init ();
 		virtual int initValues ();
 		virtual int info ();
-		virtual int setTo (float num);
-		virtual float tcOffset () {return 0.;};
+		virtual int setTo (double num);
+		virtual double tcOffset () {return 0.;};
 		virtual int isFocusing ();
 };
 
@@ -208,7 +208,7 @@ Optec::info ()
 	return Focusd::info ();
 }
 
-int Optec::setTo (float num)
+int Optec::setTo (double num)
 {
 	char command[7], rbuf[7];
 	char add = ' ';
